requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
PyMySQL==1.1.1
SQLAlchemy==2.0.0
PyJWT==2.9.0
google-auth==2.35.0
bcrypt==4.2.0
passlib==1.7.4
requests==2.32.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
cloud-sql-python-connector==1.12.1
